Initial commit
Create the first commit of my personal blog I've been working on
locally as I follow along with the curriculum on The Odin Project.
Contains the index file with all the blog posts, a secondary about
page with information about me, and a picture of a dog used in one
of the blogs.